requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.6.0
pandas==2.0.3
psycopg2-binary==2.9.7
sqlalchemy==1.4.53
